+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 Feature and bugfixes for each release.
 
+### 0.8.3
+
+* Significantly reduced vsix size
+* Refactored query parsing to handle more types of queries
+* Cleaning up memory usage after run
+
+### 0.8.2
+
+* fixed sqlite connection issue
+
 ### 0.8.0
 
 * Added a result.piton.csv to summarize the results